+++ src/components/heading/heading.scss
@@ -3,7 +3,6 @@
     font-weight: normal;
     font-size: 40px;
     margin: 0;
-    max-width: 80%;
     text-transform: uppercase;
 
     &-container {
