+++ src/components/vega-wallet-prompt/index.ts
@@ -1,0 +1,1 @@
+export * from "./vega-wallet-prompt";
