+++ src/components/vega-wallet-prompt/vega-wallet-prompt.tsx
@@ -1,0 +1,37 @@
+import React from "react";
+import { useTranslation } from "react-i18next";
+import {
+  AppStateActionType,
+  useAppState,
+  VegaKeyExtended,
+} from "../../contexts/app-state/app-state-context";
+import { useVegaUser } from "../../hooks/use-vega-user";
+
+interface VegaWalletPromptProps {
+  children: (key: VegaKeyExtended) => React.ReactElement;
+}
+
+export const VegaWalletPrompt = ({ children }: VegaWalletPromptProps) => {
+  const { t } = useTranslation();
+  const { currVegaKey } = useVegaUser();
+  const { appDispatch } = useAppState();
+
+  if (!currVegaKey) {
+    return (
+      <p>
+        <button
+          onClick={() =>
+            appDispatch({
+              type: AppStateActionType.SET_VEGA_WALLET_OVERLAY,
+              isOpen: true,
+            })
+          }
+        >
+          {t("Connect to Vega wallet")}
+        </button>
+      </p>
+    );
+  }
+
+  return children(currVegaKey);
+};
