--- src/routes/rewards/withdraw/index.ts
@@ -1,1 +1,0 @@
-export * from "./withdraw";
